requirements.txt
flask~=1.1.2
pymongo~=3.10.1
colorlog~=4.1.0
requests~=2.24.0
tencentcloud-sdk-python
cos-python-sdk-v5
qcloud-python-sts
